arm: [MVE intrinsics] rework vuninitialized
Implement vuninitialized using the new MVE builtins framework.

This patch is a bit more intrusive than other similar patches because
the two vuninitialized variants use different signatures.

To handle this situation, the patch makes the inherent shape derive
from overloaded_base instead of nonoverloaded_base, and adds a new
parameter to add_unique_function so that we can choose between adding
only the nonoverloaded version, only the overloaded ones, or both (the
default).

The patch introduces pop_and_resolve_to, which pops an argument (the
only one in this case), and calls resolve_to, like many other shapes:
this is because the overloaded version takes an argument (used to
resolve which nonoverloaded version to call), which is not present in
the nonoverloaded version prototype.

gcc/ChangeLog:
	* config/arm/arm-mve-builtins-shapes.cc (build_one): Add
	which_overload parameter.
	(inherent): Derive from overloaded_base<0>.  Add support for
	overloaded version.
	* config/arm/arm-mve-builtins-shapes.h (inherent): Update comment.
	* config/arm/arm-mve-builtins.cc (add_unique_function): Add
	support for new which_overload parameter.
	(pop_and_resolve_to): New.
	* config/arm/arm-mve-builtins.h (NONOVERLOADED_FORM)
	(OVERLOADED_FORM): New.
	(add_unique_function): Update prototype.
	(pop_and_resolve_to): New prototype.
	* config/arm/arm_mve.h (vuninitializedq): Delete.
	(vuninitializedq_u8): Delete.
	(vuninitializedq_u16): Delete.
	(vuninitializedq_u32): Delete.
	(vuninitializedq_u64): Delete.
	(vuninitializedq_s8): Delete.
	(vuninitializedq_s16): Delete.
	(vuninitializedq_s32): Delete.
	(vuninitializedq_s64): Delete.
	(vuninitializedq_f16): Delete.
	(vuninitializedq_f32): Delete.
	(__arm_vuninitializedq): Delete.