LoongArch: Implement __init_loongarch_features_resolver.
This patch can obtain the CPUCFG and HWCAP value at runtime and
extract the flag bits of features for function selection.
HWCAP is used to obtain the support of LSX and LASX because the
kernel can control the enable/disable of these two features.

Note that this requires glibc version 2.38 or higher to compile
and run.

libgcc/ChangeLog:

	* config/loongarch/t-loongarch64: Add cpuinfo.c to LIB2ADD.
	* config/loongarch/cpuinfo.c: New file.