Fix test case and add errors for when -fexternal-blas64 makse no sense.
-fexternal-blas64 requires front-end optimization to be turned on.
This patch issues a hard error if this is not the case, and also issues
an error on not-64 bit systems, where specifying it would make no sense,
and lead to errors.  Finally, this makes sure that the test is only performed
on 64-bit systems.

I tried creating test cases for the hard errors for the wrong option
combinations, but didn't succeed; see the thread on the gcc mailing
list on that topic.  These can always be added afterwards.

gcc/fortran/ChangeLog:

	PR fortran/121161
	* invoke.texi: Mention that -ffrontend-optimize is required
	for -fexternal-blas64.
	* options.cc (gfc_post_options): Fatal error if -fexternal-blas64
	is specified without -ffrontend-optimize.
	* trans-types.cc (gfc_init_kinds): Fatal error if -fexternal-blas64
	is specified on a system which does not have 64-bit ptrdiff_t.

gcc/testsuite/ChangeLog:

	PR fortran/121161
	* gfortran.dg/matmul_blas_3.f90: Add effective target lp64.