Ada: Fix internal error on use clause present in generic formal part
This is a regression present on the mainline and 15 branch: the compiler
aborts on a use clause present in the formal part of a generic unit because
of an oversight in the new inference code for generic actual parameters.

The fix also adds a missing test to Analyze_Dimension_Array_Aggregate.

gcc/ada/
	PR ada/121968
	* sem_ch12.adb (Associations.Find_Assoc): Add guard for clauses.
	* sem_dim.adb (Analyze_Dimension_Array_Aggregate): Add test for
	N_Iterated_Component_Association nodes.