vect: Extend `vec_init_loop_exit_info' to handle uncounted loops
In its current implementation, the loop vectorizer requires the main
exit be the counting IV exit. With uncounted loops we no longer need
to have any counting IV exits.  Furthermore, it is possible to have
reached this stage with malformed loops with no exits at all.

Consequently, we need an approach to handle malformed loops and some
logic to follow when choosing the main exit, when counting IV is no
longer a valid criterion.

For malformed loops, it is sufficient to return NULL, so that we can
reject such loops upon the function return.

In the case of multiple exits and no counting IV exit, we choose the
last one in the loop.  This is done so that we continue to have an
empty effective latch.

As a consequence of allowing the main exit to no longer be associated
with IV counting, the old nomenclature of `LOOP_VINFO_IV_EXIT' and
`vec_loop_iv_exit' no longer fully cover the usage of such fields and
accessors.  With that in mind, these are modified to replace "IV" for
"MAIN" for these.

gcc/ChangeLog:

	* tree-vectorizer.h (LOOP_VINFO_IV_EXIT): Replace this...
	(LOOP_VINFO_MAIN_EXIT): ...with this.
	(LOOP_VINFO_EPILOGUE_IV_EXIT): Replace this...
	(LOOP_VINFO_EPILOGUE_MAIN_EXIT): ...with this.
	(LOOP_VINFO_SCALAR_IV_EXIT): Replace this...
	(LOOP_VINFO_SCALAR_MAIN_EXIT): ...with this.
	(class _loop_vec_info): s/iv/main for `vec_loop_iv_exit',
	`vec_epilogue_loop_main_exit' and `scalar_loop_main_exit'
	class members.
	* tree-vect-data-refs.cc
	(vect_enhance_data_refs_alignment):
	s/LOOP_VINFO_IV_EXIT/LOOP_VINFO_MAIN_EXIT/.
	* tree-vect-loop-manip.cc
	(vect_set_loop_controls_directly): Likewise.
	(vect_gen_vector_loop_niters_mult_vf): Likewise.
	(vect_loop_versioning): Likewise.
	(vect_do_peeling):
	s/LOOP_VINFO_IV_EXIT/LOOP_VINFO_MAIN_EXIT/,
	s/LOOP_VINFO_SCALAR_IV_EXIT/LOOP_VINFO_SCALAR_MAIN_EXIT,
	s/LOOP_VINFO_EPILOGUE_IV_EXIT/LOOP_VINFO_EPILOGUE_MAIN_EXIT.
	* tree-vect-loop.cc (_loop_vec_info::_loop_vec_info):
	s/iv_exit/main_exit/ in initializer list.
	(vec_init_loop_exit_info): Handle multiple-exit uncounted loops.
	(vect_analyze_loop_form): Fix `failure_at' message.
	(vect_create_loop_vinfo):
	s/LOOP_VINFO_IV_EXIT/LOOP_VINFO_MAIN_EXIT/.
	(vect_create_epilog_for_reduction): Likewise.
	(vectorizable_live_operation): Likewise.
	(vect_update_ivs_after_vectorizer_for_early_breaks): Likewise.
	(vect_transform_loop):
	s/LOOP_VINFO_IV_EXIT/LOOP_VINFO_MAIN_EXIT/,
	s/LOOP_VINFO_SCALAR_IV_EXIT/LOOP_VINFO_SCALAR_MAIN_EXIT.
	* tree-vectorizer.cc (set_uid_loop_bbs):
	s/LOOP_VINFO_SCALAR_IV_EXIT/LOOP_VINFO_SCALAR_MAIN_EXIT.