fab/gimple-fold: Move removal of ASSUME internal function to gimple fold [PR121762]
This moves the removal of the ASSUME internal function to gimple fold.
The only thing is fab for -Og support until fab is removed either needs to stay
or changed over to a fold_stmt. I decided to change over to a fold_stmt for
internal function calls. I am almost positive this won't change much either way.

Bootstrapped and tested on x86_64-linux-gnu.

	PR tree-optimization/121762
gcc/ChangeLog:

	* gimple-fold.cc (gimple_fold_call): Remove ASSUME internal function
	calls when PROP_last_full_fold is set.
	* tree-ssa-ccp.cc (pass_fold_builtins::execute): Handling folding
	of all internal functions.

Signed-off-by: Andrew Pinski <andrew.pinski@oss.qualcomm.com>