ada: Catch Constraint_Errors on non-scalar streaming attributes
The specs for the streaming methods should use the first
subtype of the prefix attribute for the Item argument if the
prefix has a non-scalar type instead of the underlying type.

This will catch size errors for is smaller size are used for
the Item argument that has a constrained derived
type.

Additionally remove additional casts applied for the Input
attribute that convert the result to the prefix type.
This cast is only necessary for converting the result if
the Input attribute is used on a classwide type.

gcc/ada/ChangeLog:

	* exp_attr.adb (Expand_N_Attribute_Reference): Use the
	First_Subtype when creating the procedure for stream
	attributes.
	(Get_Array_Stream_Item_Type): New method for calculating
	the type for the Item argument for streaming methods.