c++: concepts and conversions [PR112632]
One case missed in my fix for this PR: Here we were omitting the
IMPLICIT_CONV_EXPR that expresses the conversion from int to char because
the target type was non-dependent and the argument was not type-dependent.
But we still need it if the argument is value-dependent.

	PR c++/112632

gcc/cp/ChangeLog:

	* pt.cc (convert_template_argument): Also force IMPLICIT_CONV_EXPR
	if the argument is value-dependent.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp2a/concepts-conv4.C: New test.