libstdc++: Reflect operator<< constraints in formatter for local_time.
The r16-3996-gdc78d691c5e5f7 commit (resolution of LWG4257) constrained the
operator<< for local_time, but didn't update the corresponding formatter. This
meant it didn't conditionally support formatting with an empty format spec ("{}"),
which is defined in terms of operator<<.

This patch addresses that by initializing __defSpec for the local_time
formatter in the same manner as it's done for sys_time. This functionality is
extracted to the _S_spec_for_tp function of __formatter_duration. As formatting
of local_time is defined and constrained in terms of operator<< for sys_time,
we can check the viability of the ostream operator for sys_time in both cases.

As default _M_chrono_spec may now be empty for local_time, the parse method
now checks if it was supplied in the format string, similarly to sys_time. The
condition for performing runtime check is expressed directly by checking if a
empty default is provided. This avoids the need to access the value of
__stream_insertable outside of the __defSpec computation.

As a note, despite their similar behavior, formatters sys_time and local_time
cannot be easily defined in terms of each other, as sys_time provides time zone
information while local_time does not.

libstdc++-v3/ChangeLog:

	* include/bits/chrono_io.h (__formatter_duration::_S_spec_for_tp):
	Extracted from defition of formatter<sys_time>::__defSpec.
	(formatter<chrono::sys_time<_Duration>, _CharT>::parse): Simplify
	condition in if contexpr.
	(formatter<chrono::sys_time<_Duration>, _CharT>::__stream_insertable):
	Remove.
	(formatter<chrono::sys_time<_Duration>, _CharT>::__defSpec)
	(formatter<chrono::local_time<_Duration>, _CharT>::__defSpec):
	Compute using __formatter_duration::_S_spec_for_tp.
	(forrmatter<chrono::sys_time<_Duration>, _CharT>::parse): Check if
	parse _M_chrono_spec
	* testsuite/std/time/format/empty_spec.cc: Extend tests for floating
	point and other non-streamable durations (years).

Reviewed-by: Jonathan Wakely <jwakely@redhat.com>
Signed-off-by: Tomasz Kamiński <tkaminsk@redhat.com>